Cluster init, Big.js added
The bare bones of implementing multi-threading via cluster has been
added. Child processes are being spun up that that watch a queue. The
big idea behind the queue is to take in frames and pipe them out to
ffmpeg. (NB: this won't work--I write as I rebase--because multiple
frames are being written at a time and a FIFO queue can't know whether
the frames were queued in order. In practice, they won't be.)