Frame caching, async gen/iter, and general chages
This commit is a bit of a mess. I held off committing for too long, and
it's become massive. The highlights of this commit are:

- frame caching: save each frame in an object to strip off those frames
  when piping to ffmpeg
- async generator/iterator: take the frames in the cache and pipe them
  to ffmpeg

The rest of the changes aren't worth discussing: small changes that
aren't of consequence yet. E.g., adding a big number library and some
attempts at using it and moving around code.