Redis-like server added
Engine.js is multi-threaded, but those threads don't have a central
store for memory. (NB: worker threads _do_ have shared memory, but they
weren't stable when I started this project--#29512--it's a future goal
to move this to worker threads). So, I've added a simple server for
sharing memory with redis as the inspiration. I'm not using express
because the server is so small it doesn't warrant adding a framework.