NPM scripts added, frame deletion modified
The following npm commands were added:
  - `npm generate` to generate an mp4
  - `npm server` to run the server for generation

Currently, the way an mp4 is generated is by creating each frame and
piping them to ffmpeg. To keep the memory footprint small, I'm caching
each frame as it gets generated. This allows both the right frame to be
piped at the right time (in order, since there are multiple frames being
generated at a time), and to let us only keep track of those frames that
need to be tracked (i.e., those frames that have yet to be written).
The memory footprint is kept small by deleting written frames from the
cache.

With that background, the change made to how frames are deleted in this
commit should make more sense: rather than delete them early in the
process of piping to ffmpeg, delete them only after they've been
written. It's obvious in hindsight.