Readme, mandelbrot files added; restructuring engine.js
I've moved the mandelbrot logic to its own file, as well as the math for
the mandelbrot into the new(ly committed) readme. The readme needs to be
cleaned up, but I'll do that whenever I've finished abstracting and
commenting the code.

The hope is that this cleans up the code by distinguishing the
mandelbrot logic and the logic related to multi-threading. The code
itself is still in rough-draft form; so, it'll take until its finished
to tell whether splitting up the logic makes it more readable.