Restructuring code, TODOifying
The mandelbrot set is generable using multiple cores, but the code is
atrocious. I'm starting the refactor, and with that, I've added a bunch
of TODOs and have restructured some of the code. The next steps will be
to functionify everything, abstract it, and comment it.

I've also started using a passed-in FRAME_COUNT to determine the total
number of frames to generate